fix ui layout and functions
